backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg==3.3.3
psycopg-binary==3.3.3
redis==6.4.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.38.0
whitenoise[brotli]==6.11.0
